+++ JuliaBUGS/test/runtests.jl
@@ -47,7 +47,7 @@ using SliceSampling
 
 JuliaBUGS.@bugs_primitive Beta Bernoulli Categorical Exponential Gamma InverseGamma Normal Uniform LogNormal Poisson
 JuliaBUGS.@bugs_primitive Diagonal Dirichlet LKJ MvNormal
-JuliaBUGS.@bugs_primitive censored product_distribution truncated
+JuliaBUGS.@bugs_primitive product_distribution
 JuliaBUGS.@bugs_primitive fill ones zeros
 JuliaBUGS.@bugs_primitive sum mean sqrt